make algorithms a module.
