+++ modopt/opt/algorithms/forward_backward.py
@@ -1,335 +1,13 @@
 # -*- coding: utf-8 -*-
-
-r"""OPTIMISATION ALGOTITHMS.
-
-This module contains class implementations of various optimisation algoritms.
-
-:Authors: Samuel Farrens <samuel.farrens@cea.fr>,
-         Zaccharie Ramzi <zaccharie.ramzi@cea.fr>
-
-:Notes:
-
-Input classes must have the following properties:
-
-    * **Gradient Operators**
-
-    Must have the following methods:
-
-        * ``get_grad()`` - calculate the gradient
-
-    Must have the following variables:
-
-        * ``grad`` - the gradient
-
-    * **Linear Operators**
-
-    Must have the following methods:
-
-        * ``op()`` - operator
-        * ``adj_op()`` - adjoint operator
-
-    * **Proximity Operators**
-
-    Must have the following methods:
-
-        * ``op()`` - operator
-
-The following notation is used to implement the algorithms:
-
-    * x_old is used in place of :math:`x_{n}`.
-    * x_new is used in place of :math:`x_{n+1}`.
-    * x_prox is used in place of :math:`\tilde{x}_{n+1}`.
-    * x_temp is used for intermediate operations.
-
-"""
-
-from inspect import getmro
+"""Forward-Backward Algorithms."""
 
 import numpy as np
-from progressbar import ProgressBar
 
 from modopt.base import backend
-from modopt.base.observable import MetricObserver, Observable
-from modopt.interface.errors import warn
 from modopt.opt.cost import costObj
 from modopt.opt.linear import Identity
 
-
-class SetUp(Observable):
-    r"""Algorithm Set-Up.
-
-    This class contains methods for checking the set-up of an optimisation
-    algotithm and produces warnings if they do not comply.
-
-    Parameters
-    ----------
-    metric_call_period : int, optional
-        Metric call period (default is ``5``)
-    metrics : dict, optional
-        Metrics to be used (default is ``\{\}``)
-    verbose : bool, optional
-        Option for verbose output (default is ``False``)
-    progress : bool, optional
-        Option to display progress bar (default is ``True``)
-    step_size : int, optional
-        Generic step size parameter to override default algorithm
-        parameter name (`e.g.` `step_size` will override the value set for
-        `beta_param` in `ForwardBackward`)
-    use_gpu : bool, optional
-        Option to use available GPU
-
-    """
-
-    def __init__(
-        self,
-        metric_call_period=5,
-        metrics=None,
-        verbose=False,
-        progress=True,
-        step_size=None,
-        compute_backend='numpy',
-        **dummy_kwargs,
-    ):
-
-        self.converge = False
-        self.verbose = verbose
-        self.progress = progress
-        self.metrics = metrics
-        self.step_size = step_size
-        self._op_parents = (
-            'GradParent',
-            'ProximityParent',
-            'LinearParent',
-            'costObj',
-        )
-
-        self.metric_call_period = metric_call_period
-
-        # Declaration of observers for metrics
-        super().__init__(['cv_metrics'])
-
-        for name, dic in self.metrics.items():
-            observer = MetricObserver(
-                name,
-                dic['metric'],
-                dic['mapping'],
-                dic['cst_kwargs'],
-                dic['early_stopping'],
-            )
-            self.add_observer('cv_metrics', observer)
-
-        xp, compute_backend = backend.get_backend(compute_backend)
-        self.xp = xp
-        self.compute_backend = compute_backend
-
-    @property
-    def metrics(self):
-        """Metrics."""
-        return self._metrics
-
-    @metrics.setter
-    def metrics(self, metrics):
-
-        if isinstance(metrics, type(None)):
-            self._metrics = {}
-        elif isinstance(metrics, dict):
-            self._metrics = metrics
-        else:
-            raise TypeError(
-                'Metrics must be a dictionary, not {0}.'.format(type(metrics)),
-            )
-
-    def any_convergence_flag(self):
-        """Check convergence flag.
-
-        Return if any matrices values matched the convergence criteria.
-
-        Returns
-        -------
-        bool
-            True if any convergence criteria met
-
-        """
-        return any(
-            obs.converge_flag for obs in self._observers['cv_metrics']
-        )
-
-    def copy_data(self, input_data):
-        """Copy Data.
-
-        Set directive for copying data.
-
-        Parameters
-        ----------
-        input_data : numpy.ndarray
-            Input data
-
-        Returns
-        -------
-        numpy.ndarray
-            Copy of input data
-
-        """
-        return self.xp.copy(backend.change_backend(
-            input_data,
-            self.compute_backend,
-        ))
-
-    def _check_input_data(self, input_data):
-        """Check input data type.
-
-        This method checks if the input data is a numpy array
-
-        Parameters
-        ----------
-        input_data : numpy.ndarray
-            Input data array
-
-        Raises
-        ------
-        TypeError
-            For invalid input type
-
-        """
-        if not (isinstance(input_data, (self.xp.ndarray, np.ndarray))):
-            raise TypeError(
-                'Input data must be a numpy array or backend array',
-            )
-
-    def _check_param(self, param_val):
-        """Check algorithm parameters.
-
-        This method checks if the specified algorithm parameters are floats
-
-        Parameters
-        ----------
-        param_val : float
-            Parameter value
-
-        Raises
-        ------
-        TypeError
-            For invalid input type
-
-        """
-        if not isinstance(param_val, float):
-            raise TypeError('Algorithm parameter must be a float value.')
-
-    def _check_param_update(self, param_update):
-        """Check algorithm parameter update methods.
-
-        This method checks if the specified algorithm parameters are floats
-
-        Parameters
-        ----------
-        param_update : function
-            Callable function
-
-        Raises
-        ------
-        TypeError
-            For invalid input type
-
-        """
-        param_conditions = (
-            not isinstance(param_update, type(None))
-            and not callable(param_update)
-        )
-
-        if param_conditions:
-            raise TypeError(
-                'Algorithm parameter update must be a callabale function.',
-            )
-
-    def _check_operator(self, operator):
-        """Check set-Up.
-
-        This method checks algorithm operator against the expected parent
-        classes
-
-        Parameters
-        ----------
-        operator : str
-            Algorithm operator to check
-
-        """
-        if not isinstance(operator, type(None)):
-            tree = [op_obj.__name__ for op_obj in getmro(operator.__class__)]
-
-            if not any(parent in tree for parent in self._op_parents):
-                message = '{0} does not inherit an operator parent.'
-                warn(message.format(str(operator.__class__)))
-
-    def _compute_metrics(self):
-        """Compute metrics during iteration.
-
-        This method create the args necessary for metrics computation, then
-        call the observers to compute metrics
-
-        """
-        kwargs = self.get_notify_observers_kwargs()
-        self.notify_observers('cv_metrics', **kwargs)
-
-    def _iterations(self, max_iter, progbar=None):
-        """Iterate method.
-
-        Iterate the update step of the given algorithm.
-
-        Parameters
-        ----------
-        max_iter : int
-            Maximum number of iterations
-        progbar : progressbar.ProgressBar
-            Progress bar (default is ``None``)
-
-        """
-        for idx in range(max_iter):
-            self.idx = idx
-            self._update()
-
-            # Calling metrics every metric_call_period cycle
-            # Also calculate at the end (max_iter or at convergence)
-            # We do not call metrics if metrics is empty or metric call
-            # period is None
-            if self.metrics and self.metric_call_period is not None:
-
-                metric_conditions = (
-                    self.idx % self.metric_call_period == 0
-                    or self.idx == (max_iter - 1)
-                    or self.converge,
-                )
-
-                if metric_conditions:
-                    self._compute_metrics()
-
-            if self.converge:
-                if self.verbose:
-                    print(' - Converged!')
-                break
-
-            if not isinstance(progbar, type(None)):
-                progbar.update(idx)
-
-    def _run_alg(self, max_iter):
-        """Run algorithm.
-
-        Run the update step of a given algorithm up to the maximum number of
-        iterations.
-
-        Parameters
-        ----------
-        max_iter : int
-            Maximum number of iterations
-
-        """
-        if self.progress:
-            with ProgressBar(
-                redirect_stdout=True,
-                max_value=max_iter,
-            ) as progbar:
-                self._iterations(max_iter, progbar=progbar)
-        else:
-            self._iterations(max_iter)
+from .base import SetUp
 
 
 class FISTA(object):
@@ -1086,267 +764,6 @@ class GenForwardBackward(SetUp):
             'z_new': self._z,
             'idx': self.idx,
         }
-
-    def retrieve_outputs(self):
-        """Retrieve outputs.
-
-        Declare the outputs of the algorithms as attributes: x_final,
-        y_final, metrics.
-
-        """
-        metrics = {}
-        for obs in self._observers['cv_metrics']:
-            metrics[obs.name] = obs.retrieve_metrics()
-        self.metrics = metrics
-
-
-class Condat(SetUp):
-    """Condat optimisation.
-
-    This class implements algorithm 3.1 from :cite:`condat2013`
-
-    Parameters
-    ----------
-    x : numpy.ndarray
-        Initial guess for the primal variable
-    y : numpy.ndarray
-        Initial guess for the dual variable
-    grad : class instance
-        Gradient operator class
-    prox : class instance
-        Proximity primal operator class
-    prox_dual : class instance
-        Proximity dual operator class
-    linear : class instance, optional
-        Linear operator class (default is ``None``)
-    cost : class or str, optional
-        Cost function class (default is 'auto'); Use 'auto' to automatically
-        generate a costObj instance
-    reweight : class instance, optional
-        Reweighting class
-    rho : float, optional
-        Relaxation parameter (default is ``0.5``)
-    sigma : float, optional
-        Proximal dual parameter (default is ``1.0``)
-    tau : float, optional
-        Proximal primal paramater (default is ``1.0``)
-    rho_update : function, optional
-        Relaxation parameter update method (default is ``None``)
-    sigma_update : function, optional
-        Proximal dual parameter update method (default is ``None``)
-    tau_update : function, optional
-        Proximal primal parameter update method (default is ``None``)
-    auto_iterate : bool, optional
-        Option to automatically begin iterations upon initialisation (default
-        is ``True``)
-    max_iter : int, optional
-        Maximum number of iterations (default is ``150``)
-    n_rewightings : int, optional
-        Number of reweightings to perform (default is ``1``)
-
-    Notes
-    -----
-    The `tau_param` can also be set using the keyword `step_size`, which will
-    override the value of `tau_param`.
-
-    See Also
-    --------
-    SetUp : parent class
-
-    """
-
-    def __init__(
-        self,
-        x,
-        y,
-        grad,
-        prox,
-        prox_dual,
-        linear=None,
-        cost='auto',
-        reweight=None,
-        rho=0.5,
-        sigma=1.0,
-        tau=1.0,
-        rho_update=None,
-        sigma_update=None,
-        tau_update=None,
-        auto_iterate=True,
-        max_iter=150,
-        n_rewightings=1,
-        metric_call_period=5,
-        metrics=None,
-        **kwargs,
-    ):
-
-        # Set default algorithm properties
-        super().__init__(
-            metric_call_period=metric_call_period,
-            metrics=metrics,
-            **kwargs,
-        )
-
-        # Set the initial variable values
-        for input_data in (x, y):
-            self._check_input_data(input_data)
-
-        self._x_old = self.xp.copy(x)
-        self._y_old = self.xp.copy(y)
-
-        # Set the algorithm operators
-        for operator in (grad, prox, prox_dual, linear, cost):
-            self._check_operator(operator)
-
-        self._grad = grad
-        self._prox = prox
-        self._prox_dual = prox_dual
-        self._reweight = reweight
-        if isinstance(linear, type(None)):
-            self._linear = Identity()
-        else:
-            self._linear = linear
-        if cost == 'auto':
-            self._cost_func = costObj([
-                self._grad,
-                self._prox,
-                self._prox_dual,
-            ])
-        else:
-            self._cost_func = cost
-
-        # Set the algorithm parameters
-        for param_val in (rho, sigma, tau):
-            self._check_param(param_val)
-
-        self._rho = rho
-        self._sigma = sigma
-        self._tau = self.step_size or tau
-
-        # Set the algorithm parameter update methods
-        for param_update in (rho_update, sigma_update, tau_update):
-            self._check_param_update(param_update)
-
-        self._rho_update = rho_update
-        self._sigma_update = sigma_update
-        self._tau_update = tau_update
-
-        # Automatically run the algorithm
-        if auto_iterate:
-            self.iterate(max_iter=max_iter, n_rewightings=n_rewightings)
-
-    def _update_param(self):
-        """Update parameters.
-
-        This method updates the values of the algorthm parameters with the
-        methods provided
-
-        """
-        # Update relaxation parameter.
-        if not isinstance(self._rho_update, type(None)):
-            self._rho = self._rho_update(self._rho)
-
-        # Update proximal dual parameter.
-        if not isinstance(self._sigma_update, type(None)):
-            self._sigma = self._sigma_update(self._sigma)
-
-        # Update proximal primal parameter.
-        if not isinstance(self._tau_update, type(None)):
-            self._tau = self._tau_update(self._tau)
-
-    def _update(self):
-        """Update.
-
-        This method updates the current reconstruction
-
-        Notes
-        -----
-        Implements equation 9 (algorithm 3.1) from :cite:`condat2013`
-
-        - primal proximity operator set up for positivity constraint
-
-        """
-        # Step 1 from eq.9.
-        self._grad.get_grad(self._x_old)
-
-        x_prox = self._prox.op(
-            self._x_old - self._tau * self._grad.grad - self._tau
-            * self._linear.adj_op(self._y_old),
-        )
-
-        # Step 2 from eq.9.
-        y_temp = (
-            self._y_old + self._sigma
-            * self._linear.op(2 * x_prox - self._x_old)
-        )
-
-        y_prox = (
-            y_temp - self._sigma
-            * self._prox_dual.op(
-                y_temp / self._sigma,
-                extra_factor=(1.0 / self._sigma),
-            )
-        )
-
-        # Step 3 from eq.9.
-        self._x_new = self._rho * x_prox + (1 - self._rho) * self._x_old
-        self._y_new = self._rho * y_prox + (1 - self._rho) * self._y_old
-
-        del x_prox, y_prox, y_temp
-
-        # Update old values for next iteration.
-        self.xp.copyto(self._x_old, self._x_new)
-        self.xp.copyto(self._y_old, self._y_new)
-
-        # Update parameter values for next iteration.
-        self._update_param()
-
-        # Test cost function for convergence.
-        if self._cost_func:
-            self.converge = (
-                self.any_convergence_flag()
-                or self._cost_func.get_cost(self._x_new, self._y_new)
-            )
-
-    def iterate(self, max_iter=150, n_rewightings=1):
-        """Iterate.
-
-        This method calls update until either convergence criteria is met or
-        the maximum number of iterations is reached
-
-        Parameters
-        ----------
-        max_iter : int, optional
-            Maximum number of iterations (default is ``150``)
-        n_rewightings : int, optional
-            Number of reweightings to perform (default is ``1``)
-
-        """
-        self._run_alg(max_iter)
-
-        if not isinstance(self._reweight, type(None)):
-            for _ in range(n_rewightings):
-                self._reweight.reweight(self._linear.op(self._x_new))
-                self._run_alg(max_iter)
-
-        # retrieve metrics results
-        self.retrieve_outputs()
-        # rename outputs as attributes
-        self.x_final = self._x_new
-        self.y_final = self._y_new
-
-    def get_notify_observers_kwargs(self):
-        """Notify observers.
-
-        Return the mapping between the metrics call and the iterated
-        variables.
-
-        Returns
-        -------
-        notify_observers_kwargs : dict,
-           The mapping between the iterated variables
-
-        """
-        return {'x_new': self._x_new, 'y_new': self._y_new, 'idx': self.idx}
 
     def retrieve_outputs(self):
         """Retrieve outputs.
